Move styled to parent dir
