+++ src/main/styled.ts
@@ -1,4 +1,4 @@
 import styled, { CreateStyled } from '@emotion/styled'
-import { Theme } from '../themes'
+import { Theme } from './themes'
 
 export default styled as CreateStyled<Theme>
